chore: fully port multiaddr test to cmocka
